5 th commit assignment 4 (part 4)
this section is the Class Shape, which is the base class with properties for position and velocity. Ball extends Shape and adds properties and methods for drawing, updating, and detecting collisions with other balls.
EvilCircle extends Shape and adds methods for drawing, keeping within bounds, and detecting collisions with balls.