assignment 5 commit 4
This section handles the functionality for showing and hiding the comment section as well as adding new comments. When the user clicks on the "Show/Hide comments" button, it toggles the visibility of the comment wrapper. Additionally, when the user submits the comment form, a new comment is added to the list of comments, and the form is reset